Move show snack fun from router to ext
