+++ app/src/main/java/io/vasilenko/otus/moviesearcher/presentation/MoviesActivity.kt
@@ -5,10 +5,8 @@ import android.view.MenuItem
 import android.view.View
 import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
-import com.google.android.material.snackbar.Snackbar
 import io.vasilenko.otus.moviesearcher.MovieSearcherApp
 import io.vasilenko.otus.moviesearcher.R
-import io.vasilenko.otus.moviesearcher.presentation.common.MessageBundle
 import io.vasilenko.otus.moviesearcher.presentation.common.QuitDialog
 import io.vasilenko.otus.moviesearcher.presentation.navigation.MoviesRouter
 import io.vasilenko.otus.moviesearcher.presentation.navigation.MoviesRouterHandler
@@ -75,18 +73,6 @@ class MoviesActivity : AppCompatActivity(), MoviesView, MoviesRouterHandler {
         if (addToBackStack) transaction.addToBackStack(fragment.javaClass.name)
         if (!showNavBar) moviesBottomNavigation.visibility = View.GONE
         transaction.commit()
-    }
-
-    override fun onMessage(messageBundle: MessageBundle) {
-        val snackbar = Snackbar.make(
-            requireViewById(R.id.moviesContainer),
-            messageBundle.text,
-            Snackbar.LENGTH_SHORT
-        )
-        messageBundle.action?.let {
-            snackbar.setAction(messageBundle.action.name, messageBundle.action.listener).show()
-        }
-        snackbar.show()
     }
 
     private fun navBarListener(item: MenuItem): Boolean {
