+++ app/src/main/java/io/vasilenko/otus/moviesearcher/presentation/navigation/MoviesRouter.kt
@@ -1,13 +1,10 @@
 package io.vasilenko.otus.moviesearcher.presentation.navigation
 
 import androidx.fragment.app.Fragment
-import io.vasilenko.otus.moviesearcher.presentation.common.MessageBundle
 
 interface MoviesRouter {
 
     fun setRouterHandler(routerHandler: MoviesRouterHandler)
 
     fun onOpenFragment(fragment: Fragment, addToBackStack: Boolean, showNavBar: Boolean)
-
-    fun onMessage(messageBundle: MessageBundle)
 }
