+++ app/src/main/java/io/vasilenko/otus/moviesearcher/presentation/navigation/impl/MoviesRouterImpl.kt
@@ -1,7 +1,6 @@
 package io.vasilenko.otus.moviesearcher.presentation.navigation.impl
 
 import androidx.fragment.app.Fragment
-import io.vasilenko.otus.moviesearcher.presentation.common.MessageBundle
 import io.vasilenko.otus.moviesearcher.presentation.navigation.MoviesRouter
 import io.vasilenko.otus.moviesearcher.presentation.navigation.MoviesRouterHandler
 
@@ -15,9 +14,5 @@ class MoviesRouterImpl : MoviesRouter {
 
     override fun onOpenFragment(fragment: Fragment, addToBackStack: Boolean, showNavBar: Boolean) {
         routerHandler?.onOpenFragment(fragment, addToBackStack, showNavBar)
-    }
-
-    override fun onMessage(messageBundle: MessageBundle) {
-        routerHandler?.onMessage(messageBundle)
     }
 }
